server: start work on server & web ui
The interface for controlling lagopus will be an HTTP API server along
with a web client. These are implemented in the same app for simplicity.

lagopus.py has been renamed to lagopus-client.py in anticipation of it
becoming a thin client of this server, though it isn't yet.

Current tech is a flask app serving a JSON API and a web app that uses
that API. The same app will interface directly with the Kubernetes API
to create and query resources and make them available to external
clients.

All of this is built in a container so people don't have to deal with
the hellish js ecosystem.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>