lagopus-fuzzer: perform post-fuzzing processing
- Verify and minimize crashing test cases
- Analyze crashes with the `exploitable` gdb plugin, and store the
  results in `crashes/crashes.db` in the job results zip
- Minimize corpus results

All courtesy of `afl-collect`.

Because the above steps take variable time to run (especially
minimization, which may take a really long time) the deadline on the k8s
Job resource has been removed, so it's now really important to make sure
the in-container deadline is respected so we don't end up fuzzing
forever.

This commit also breaks libFuzzer support even more than it probably
already is. That will need to be revisited since the tooling for
libFuzzer isn't anywhere near as mature so verification and crash
analysis will likely be harder to implement.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>