lagopus-fuzzer: fix job name jq extraction
jq was returning the afl-multicore session name with "quotes", which
afl-multikill doesn't like. Returning the value without qutoes allows
afl-multikill to find the job PID and kill it.

However, this also seems to be PID 1, and it also doesn't seem to be
working even with the patch, but this is at least more correct.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>