server, fuzzer: add more stats, combine graphs
Add a few more stats from AFL, and combine the many graphs into two
graphs - one for performance monitoring that only shows execs / sec, and
the other that shows path statistics. Crashes and hangs aren't shown and
should be shown probably in a gauge or something instead of a time
series graph.

Also, hyperlink job names in crashes table to the job page.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>