server: update status of deleted jobs
Before updating the status field on jobs in the database, set all the
jobs that aren't in "Complete" to "Unknown". This way when the real
status from k8s is used to update those jobs, jobs that have been
deleted and were never marked completed get transitioned to "Unknown"
status.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>