server: downsample stats data to 1 minute avg
Long jobs can produce so much stat data that it gets to several mb
returned in json via the API. Nothing needs this resolution, indeed when
the # data points gets in the 10s of thousands chartjs loses its cool
and page perf drops to nothing. This patch is a global hammer to return
only the averages of data over 1m time periods, which fixes page perf
and should be fine for future components.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>