server: merge jobs page into dashboard
They both have basically the same data; jobs page is still around, but
now only used to show a particular job's details. Table code etc is
still there and the page can still be accessed but it's not shown in the
sidebar anymore. New job button moves to dashboard.

Also tweaks the code for the summary boxes above the jobs table on the
dashboard.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>