server: add some job validation
Was getting a bit annoying when I screwed up the job files and there was
no feedback about what was wrong. This adds some simple server-side
checks on the uploaded zip and reports any issues back to the frontend.
The frontend displays these.

This breaks page refresh on successful job upload, something to do with
the completion callback for jquery ajax calls being present that
overrides the reload in the success callback.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>