fuzzer: handle non-reproducing libFuzzer crashes
Sometimes, crashes found by libFuzzer depends on program state
accumulated during previous test case runs, for example, a crash that
only happens on the 100th test case. From libFuzzer's point of view, it
was the 100th test case that crashed the program, but in reality only
feeding that one case won't reproduce the crash because it depends on
state accumulated during the previous 99 runs. In this case, our crash
analysis workflow won't reproduce the crash because it runs the target
program against only the saved test case. This can be annoying because
the crash will be logged in the database but there won't be a stack
trace or any way to reproduce it locally without rerunning the entire
corpus manually, which defeats the point of automated infrastructure.

In this scenario we need to go look at the fuzzing logs and retrieve the
stack trace from there. For now we select the last 300 lines of the log
file, hoping that the stack trace won't be more than 300 lines long.
Since the analysis tooling still returns the complete output as the
stack trace when analysis is done, we will end up saving extraneous
program output as part of the trace. This needs to be revisited and the
saved trace needs to have extraneous output filtered out.

In any case, this patch allows the user to still see the stack trace
even if it won't reproduce from a single input. In this case the return
code is set to 101, but this should probably be encapsulated in a
separate flag field in the database indicating whether the crash is
reproducible from the provided sample.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>